tdf#127205 split Desktop::terminate process
Trying to somehow keep stuff correctly alive, truned out to be
rather futile. Originally I tried the same way Caolan did in the
attached patch to tdf#88985, when he finally settled on adding a
special terminate listener.

But this is not enough to handle in-process scripted extensions,
like Java, Python or even Basic macros themself.

So this separates the module shutdown and SfxApplication cleanup
from the rest of the terminate call.

Change-Id: Ice59816080c922a17511b68afe59348a662600c9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88835
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>