tdf#80194 UI: revert subscript DFLT_ESC_SUB to 8% (from 33%)
Originally this was at 8%, but long ago for some reason that
I couldn't find, it was changed to match superscript's 33%.
This primarily affects UI for editeng.

33% was completely wrong. It puts the subscript WAY
too low. The font's descent is only 20% of the total font height,
so to lower the character by 33% pushes it below the line
instead of towards the bottom of the line.

Many export fixes that match the formula which calculated the
values for DFLT_ESC_AUTO_* were made to LO6.4 for tdf#127316.

Changing this default suprisingly has almost no effect.
Import and export do not depend on a specific escapement
to determine whether to treat it as automatic or not.
That is generally handled with special keywords,
or in the case of RTF with a +1 on the nPROP value.
Writerfilter's use is in a never-to-be-encountered
failsafe edge-case. HTML's import is affected,
but for the good, so I made it into a unit test.

Primarily it affects the UI dialog - suggesting
8% when automatic checkbox is cleared, and the biggest
impact comes with the toolbar button for Draw/Impress
which don't set "automatic" mode, so they just take
the DFLT_ESC_SUB value. Looks much better now.
And that was the heart of the bug report - that
editeng defaults for subscript were way off.

Change-Id: I6769072d483467e86fea82dfc534eb5e04802491
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88910
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>