tdf#130274 sw_redlinehide: fix ChgAutoCorrWord() if replaced text ...
... is fully deleted.

This crashes on English text with 'i'->'I' substitution.

The assumption that there's some part of the found text that is not
deleted is only correct for the case where tracked changes are hidden,
but not when they are shown.

(regression from c8d2fc405c52805da5fa27337afdddb9f93999e0)

Change-Id: I0e81494659ea7e187101a703f64483dd68c73d70
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/89151
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>