Make headings navigation view update after delete chapter undo
This is a partial revert of commit
edd16481a0ac3760dd04b0c5324d4dd53715cf91 I did to make multiselected
outline moves keep selections in headings navigation view after move. It
seems this part of the patch is not required and causes undo of chapter
delete not to be reflected in the headings navigation view if it has
focus when undo is done.

Change-Id: I40dbe4aba83222e7d0c373e043bfb0e3099e7a61
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/89288
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>