Remove unused lokit-only selection properties in..
CID string for charts. This is messing up the parsing of
CID string in ObjectIdentifier::getObjectType() and elsewhere
where parsing CID is attempted, which inturn causes at least
the sidebar's chart line properties listboxes not getting
updated.

The selection properties insertion in chart CID was introduced in commit

8340210a207915597c4baa8023903355f8e55145

but these are not even used in online as of present.

Change-Id: Ic8040d84e1ad16f182fc7bdc3a9f533ad9664458
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/89300
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/89697
Tested-by: Jenkins