Revert "jsdialog: Refresh dialog after action"
This change makes some cypress tests to fail, the
mobile wizard does not open or opens with an empty content.
It seems it causes an issue when mobile wizard is opened/closed
subsequently. One test case does not make this visible, but
it needs to run  more test cases subsequently.

This reverts commit f0aede4515ee20183de61d6a1cad83a82372ebf9.

Change-Id: Iccb6bb3d77270340ac9b36a68cae749098ddaa3a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/89502
Tested-by: Tamás Zolnai <tamas.zolnai@collabora.com>
Reviewed-by: Tamás Zolnai <tamas.zolnai@collabora.com>