repack the boost tarball to be smaller and faster
The tarball is almost 100MiB and it unpacks to roughly 0.75GiB,
because it contains tons of generated html docs and other stuff that
is not needed for building. Unpacking it on Cygwin takes several
minutes (done in parallel, but still) and even on Linux the unpacking
takes longer than the actual compilation.

Change-Id: Ied9f3059530f4e9a856170ec3fcc304631ff8e0a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/89825
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>