tdf#130680 sw_redlinehide: fix crash in SwUndoDelete
The problem is that the code that adjusts pLastNode in
SwContentNode::DelFrames() assumes that the nodes are deleted forward
but MoveNode() runs backward so the only solution is to put pLastNode
temporarily onto some node without checking if it has been fully
deleted, while skipping over tables and sections.

At the end of MoveNode() it must be on a valid (not fully deleted) node
again.

(regression from c6a931f0a054062c3eee4849b736aefb10a58956)

Change-Id: If54c01f885e0c040aed9068823deae7dd167fd50
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/89896
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>