sw: DOCX export: fix bookmark inside sdtDropDown
Word refuses to open a document that has a w:bookmarkEnd inside
w:sdtContent but with no text content following it.

It turns out that the bookmark position is wrong anyway, it should end
before the text according to Writer's model.

It shouldn't make a difference whether the end is inside the sdtContent
or preceding the SDT, so write the text content of the SDT from the
EndField_Impl().

Another idea would be to move the writing of bookmarks in EndRun()
before the StartField_Impl() but who knows what that would break.

(regression from c467a6edf5c8f98c70943d7a3d8fd4ca617b4006)

Change-Id: I476c0829814b061d80811cc6817923ee06013a26
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/90100
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>