tdf#123116 sw layout: don't "page-break" before oversized row
followup to LO 7.0 commit f8c91a644085327e4eddfddd670c0bd71efa6bb6

THIS PATCH INTENTIONALLY BREAKS COMPATIBILITY WITH MSO.

In the rather rare case of a table row that is larger than
the page, but is nevertheless set to not allow splitting,
historically this oversized row has started on a new page
and disappeared into the bottom of the page.

The commit mentioned above allowed the row to spill over
onto the next page, but kept the "page break" feature for
maximum compatibility and interoperability with MSWord.

However, all of that seems completely senseless, so
the most natural thing for the document to do would be
to completely ignore the do-not-split-row setting and let
everything flow naturally, without leaving a potentially
large gap in order to start on a new page.
Just let the oversized row stay attached to the prior row.

The cases are rare enough that interoperability isn't important.
In any case, for .doc documents authored by LibreOffice,
Word is still going to try to squeeze it all on one page anyway,
so even if we do leave a gap and start a new page,
there will still be a difference.

Now that compatibilityMode=15 for new .docx files,
MSO almost perfectly matches the behaviour prior to this patch,
so this is the one case where users might raise a complaint.

The prior patch was committed first just to prove that
we could fallback to a more compatible state in case
this radical departure from the norm is unacceptable.

Change-Id: I47c78526d4c4fda2c48e38fb64788bafbc06f2c8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/90131
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>