tdf#131071 writerfilter: tableStyle uses DocDefaults
It is kinda surprising to me, but the tableStyle actually
includes DocDefaults in its style properties.

You would expect that to be irrelevant, because if even the
default style specifies a setting, it overrides the table style.
Therefore they would either both match DocDefaults, or the
style would have specified a value (thus overriding tablestyle).

But in the specific default case of
overrideTableStyleSizeAndJustification == false,
a left alignment or a fontsize of 11 or 12 are overridden
by the table style.

Perhaps this should ONLY import those two settings in
DocDefaults, but obviously the concept is there, so
it shouldn't hurt to import the whole thing as that
likely is the correct response for other situations
as well.

Change-Id: Ie2494855fa37c9a21d27d0e15f4da1fa0a216be6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/90558
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: László Németh <nemeth@numbertext.org>