nss: fix this up to work with Online again
Regression from commit 3575d9df284050e29e2ff377a944cdbe6840a2b7 (build
nss using their new build system (gyp/ninja-based), 2020-02-19), the
problem was that for some reason nss was pre-loaded in the past in some
magic way, but after that commit we need to do this explicitly.

Additionally, if the core.git build is a debug one, we now get an nss
assert that checks if there was a fork between loading and using nss. I
don't think that was enabled before, so just patch that out to fix the
build breakage on the online.git side.

online.git's unit-password-protected now passes again with this.

Change-Id: I56c0019c446b4cabbb73141bbff80e8d8ffdb54c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91068
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>