tdf#80635 ww8import:1 take float table CENTER to inlined table
Since SW has problems with floating tables, any unnecessarily
floated tables are converted to inline. However, the
positioning of the tables is mostly lost. So the choice is
either to refuse to inline a strangely positioned table
or else to transfer the positioning onto the inlined table.

Transferring is easier said than done. But for centered tables,
it is easy to copy that.

Change-Id: I8a6deaa63c45277c30ffa12613c8ed125a45d607
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91188
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>