tdf#131593 ScOutputData::GetMergeOrigin: Handle cells outside own area
Rendering the bugdoc had regressed at commit
5c096eeb9593ee2ec44585d263ae16e9cc223833
("LOK: fast tile rendering (text only)", 2016-06-04),
which had changed the area covered by the ScTableInfo
in use from the document's whole rendering area to that
covered by the bounding rows and columns for the tile
being rendered.

For the bug doc, it was e.g. no longer detected that cell
A10 (nX=0, nY=9) is merged to A1 when processing the
"second tile in the second row of tiles" in
ScOutputData::DrawEdit.

Change-Id: I60c3ad43ad816d2186547486116f63e27ed3a166
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91094
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>