tdf#132460 Fix feedback-loop during WebDAV lock refresh
Prevent a feedback-loop, where LibreOffice would request a shorter timeout on
each refresh.

WebDAV servers do not return the lock-timeout the lock was set up with; they
have to return the remaining time. There are WebDAV servers that return a lower
timeout. Two examples [2] [3]. Other servers probably reuse the same timestamp
for the whole request and are therefore not subject to that problem. (This might
even be incorrect if the request takes very long.)

Updating the lock-timeout with the value returned by the server decreases the
timeout until it reaches zero.

LibreOffice request               SERVER response
LOCK(180)         ->              LOCK(60)
LOCK(60)          ->              LOCK(59)
LOCK(59)          ->              LOCK(58)
...
0: no LOCK header ->              FAIL

If we do not update the timeout in NeonSession::LOCK() only the initial setup
updates the timeout.

LibreOffice request               SERVER response
LOCK(180)        ->               LOCK(60)
LOCK(60)         ->               LOCK(59)
LOCK(60)         ->               LOCK(59)
...

It is essential that lastChanceToSendRefreshRequest uses the timeout returned by
the server; after it calculated the deadline, we reset the timeout.

The maintainer of neon confirmed that the timeout should stay the same after the
initial setup. [4].

[1] https://tools.ietf.org/html/rfc4918#section-6.6
[2] https://www.alfresco.com/
[3] https://github.com/mar10/wsgidav
[4] https://github.com/notroj/neon/issues/12

Change-Id: Ie76338f7a88f41f47569a62ea6efec8c6f646f50
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/92981
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>