tdf#125268 officecfg: export LO character background as MSO shading by default
This restores commit 0352402639c5af412a84eda9e513291ca7050f29
(tdf#125268: Export LO character background as MSO shading by default.,
2020-03-23).

As per ESC discussion, see
<https://lists.freedesktop.org/archives/libreoffice/2020-April/084813.html>
for the details. Summary: the benefit is that this way we don't loose
data, and the cost is that LO's highlighting is called shading in MSO
now.

The minutes has several good ideas how to improve this further in
follow-up commits.

Change-Id: I12c76a7c273178d284b84c6c23e9348feb316127
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91611
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins